+++ hyperts/tests/hyperts_test.py
@@ -4,9 +4,9 @@ from hypernets.core.search_space import HyperSpace, Choice
 from hypernets.searchers.random_searcher import RandomSearcher
 
 from hyperts.hyper_ts import HyperTS
-from hyperts.utils import consts, data_ops as dp
+from hyperts.utils import consts, toolbox as dp
 from hyperts.datasets import load_random_univariate_forecast_dataset
-from hyperts.wrappers.stats_wrappers import SimpleTSEstimator, ProphetWrapper
+from hyperts.framework.wrappers.stats_wrappers import SimpleTSEstimator, ProphetWrapper
 
 class Test_HyperTS():
