+++ hyperts/tests/task_test.py
@@ -3,14 +3,14 @@
 from hypernets.core.callbacks import SummaryCallback
 from hypernets.core.searcher import OptimizeDirection
 from hypernets.searchers.random_searcher import RandomSearcher
-from hyperts.experiment import TSExperiment
+from hyperts.framework.compete import TSCompeteExperiment
 from hyperts.hyper_ts import HyperTS
-from hyperts.search_space import search_space_univariate_forecast_generator, search_space_multivariate_forecast_generator, \
+from hyperts.micro_search_space import search_space_univariate_forecast_generator, search_space_multivariate_forecast_generator, \
     space_classification_classification
 
 from sktime.datasets import load_arrow_head
 from hyperts.datasets import load_random_univariate_forecast_dataset, load_random_multivariate_forecast_dataset
-from hyperts.utils.data_ops import random_train_test_split, temporal_train_test_split
+from hyperts.utils.toolbox import random_train_test_split, temporal_train_test_split
 
 class Test_Task():
 
@@ -21,7 +21,7 @@ class Test_Task():
         rs = RandomSearcher(search_space_univariate_forecast_generator(covariate=['id'], timestamp='ds'), optimize_direction=OptimizeDirection.Minimize)
         hyper_model = HyperTS(rs, task='univariable-forecast', reward_metric='rmse', callbacks=[SummaryCallback()])
 
-        exp = TSExperiment(hyper_model, X_train, y_train, X_eval=X_test, y_eval=y_test, timestamp_col='ds', covariate_cols=['id'])
+        exp = TSCompeteExperiment(hyper_model, X_train, y_train, X_eval=X_test, y_eval=y_test, timestamp_col='ds', covariate_cols=['id'])
         pipeline_model = exp.run(max_trials=3)
 
         y_pred = pipeline_model.predict(X_test)
@@ -35,7 +35,7 @@ class Test_Task():
         rs = RandomSearcher(search_space_multivariate_forecast_generator(timestamp='ds'), optimize_direction=OptimizeDirection.Minimize)
         hyper_model = HyperTS(rs, task='multivariable-forecast', reward_metric='rmse', callbacks=[SummaryCallback()])
 
-        exp = TSExperiment(hyper_model, X_train, y_train, X_eval=X_test, y_eval=y_test, timestamp_col='ds')
+        exp = TSCompeteExperiment(hyper_model, X_train, y_train, X_eval=X_test, y_eval=y_test, timestamp_col='ds')
         pipeline_model = exp.run(max_trials=3)
 
         y_pred = pipeline_model.predict(X_test)
@@ -49,7 +49,7 @@ class Test_Task():
         rs = RandomSearcher(space_classification_classification, optimize_direction=OptimizeDirection.Maximize)
         hyper_model = HyperTS(rs, task='multiclass-classification', reward_metric='accuracy', callbacks=[SummaryCallback()])
 
-        exp = TSExperiment(hyper_model, X_train, y_train, X_eval=X_test, y_eval=y_test)
+        exp = TSCompeteExperiment(hyper_model, X_train, y_train, X_eval=X_test, y_eval=y_test)
         pipeline_model = exp.run(max_trials=3)
 
         y_pred = pipeline_model.predict(X_test)
